pyproject.toml
[project]
name = "workingagent"
version = "0.1.0"
description = "RAG Chatbot with FastAPI, Agno, and NiceGUI"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "agno>=0.1.0",
    "nicegui>=2.0.0",
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.9",
    "pypdf>=4.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-check>=2.0.0",
    "ruff>=0.6.0",
    "black>=24.0.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true

